+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Repls can now be stopped by `^C` signal.
  - `Noop` and `AsyncStopToken`stop tokens.
  - `StatefulListener`.
+ - Emit not only errors but also warnings and general information from teloxide, when set up by `enable_logging!`.
 
 ### Fixed